api/requirements.txt

# fastapi==0.110.0
# uvicorn==0.29.0
# python-multipart==0.0.9
# pillow==10.2.0
# numpy==1.26.4
# matplotlib==3.8.2

# tensorflow==2.10.0
# # tensorflow-serving-api==2.10.0



fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pillow==10.2.0
numpy==1.26.4
matplotlib==3.8.2

tensorflow==2.15.0  # Upgrade if compatible
# Uncomment if needed for TensorFlow Serving
# tensorflow-serving-api==2.15.0
